+++ src/redux/actions/provider-packages/get-provider-packages.js
@@ -1,8 +1,0 @@
-export const GET_PROVIDER_PACKAGES = 'GET_PROVIDER_PACKAGES';
-
-export function getProviderPackages(payload) {
-  return {
-    type: GET_PROVIDER_PACKAGES,
-    payload,
-  };
-}